+++ packages/std/src/query/mod.rs
@@ -21,7 +21,7 @@ pub use staking::{
 };
 #[cfg(feature = "stargate")]
 pub use stargate::StargateResponse;
-pub use wasm::WasmQuery;
+pub use wasm::{ContractInfoResponse, WasmQuery};
 
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]